+++ protected/controllers/AdminLinkPrefixController.php
@@ -42,6 +42,7 @@ class AdminLinkPrefixController extends Controller
      */
     public function actionView($id)
     {
+		$this->layout = 'new_datasetpage';
         $this->render('view',array(
             'model'=>$this->loadModel($id),
         ));
@@ -67,6 +68,7 @@ class AdminLinkPrefixController extends Controller
                 $this->redirect(array('view','id'=>$model->id));
         }
 
+		$this->layout = 'new_datasetpage';
         $this->render('create',array(
             'model'=>$model,
         ));
@@ -91,6 +93,7 @@ class AdminLinkPrefixController extends Controller
                 $this->redirect(array('view','id'=>$model->id));
         }
 
+		$this->layout = 'new_datasetpage';
         $this->render('update',array(
             'model'=>$model,
         ));
@@ -137,6 +140,8 @@ class AdminLinkPrefixController extends Controller
         if(isset($_GET['Prefix']))
             $model->setAttributes($_GET['Prefix']);
 
+        $this->layout = 'new_main';
+        $this->loadBaBbqPolyfills = true;
         $this->render('admin',array(
             'model'=>$model,
         ));