+++ protected/views/adminManuscript/update.php
@@ -1,18 +1,14 @@
-<?php
-$this->breadcrumbs=array(
-	'Manuscripts'=>array('index'),
-	$model->id=>array('view','id'=>$model->id),
-	'Update',
-);
+<div class="container">
+	<?php
+	$this->widget('TitleBreadcrumb', [
+		'pageTitle' => 'Update Manuscript ' . $model->id,
+		'breadcrumbItems' => [
+			['label' => 'Admin', 'href' => '/site/admin'],
+			['label' => 'Manage', 'href' => '/adminManuscript/admin'],
+			['isActive' => true, 'label' => 'Update'],
+		]
+	]);
+	?>
 
-$this->menu=array(
-	array('label'=>'List Manuscript', 'url'=>array('index')),
-	array('label'=>'Create Manuscript', 'url'=>array('create')),
-	array('label'=>'View Manuscript', 'url'=>array('view', 'id'=>$model->id)),
-	array('label'=>'Manage Manuscript', 'url'=>array('admin')),
-);
-?>
-
-<h1>Update Manuscript <?php echo $model->id; ?></h1>
-
-<?php echo $this->renderPartial('_form', array('model'=>$model)); ?>
+	<?php echo $this->renderPartial('_form', array('model' => $model)); ?>
+</div>