+++ protected/views/adminManuscript/view.php
@@ -1,26 +1,26 @@
-<?php
-$this->breadcrumbs=array(
-	'Manuscripts'=>array('index'),
-	$model->id,
-);
+<div class="container">
+	<?php
+	$this->widget('TitleBreadcrumb', [
+		'pageTitle' => 'View Manuscript #' . $model->id,
+		'breadcrumbItems' => [
+			['label' => 'Admin', 'href' => '/site/admin'],
+			['label' => 'Manage', 'href' => '/adminManuscript/admin'],
+			['isActive' => true, 'label' => 'View'],
+		]
+	]);
+	?>
 
-$this->menu=array(
-	array('label'=>'List Manuscript', 'url'=>array('index')),
-	array('label'=>'Create Manuscript', 'url'=>array('create')),
-	array('label'=>'Update Manuscript', 'url'=>array('update', 'id'=>$model->id)),
-	array('label'=>'Delete Manuscript', 'url'=>'#', 'linkOptions'=>array('submit'=>array('delete','id'=>$model->id),'confirm'=>'Are you sure you want to delete this item?')),
-	array('label'=>'Manage Manuscript', 'url'=>array('admin')),
-);
-?>
+	<?php $this->widget('zii.widgets.CDetailView', array(
+		'data' => $model,
+		'attributes' => array(
+			'id',
+			'identifier',
+			'pmid',
+			'dataset_id',
+		),
+		'htmlOptions' => array('class' => 'table table-striped table-bordered dataset-view-table'),
+		'itemCssClass' => array('odd', 'even'),
+		'itemTemplate' => '<tr class="{class}"><th scope="row">{label}</th><td>{value}</td></tr>'
+	)); ?>
 
-<h1>View Manuscript #<?php echo $model->id; ?></h1>
-
-<?php $this->widget('zii.widgets.CDetailView', array(
-	'data'=>$model,
-	'attributes'=>array(
-		'id',
-		'identifier',
-		'pmid',
-		'dataset_id',
-	),
-)); ?>
+</div>