+++ protected/views/adminSample/view.php
@@ -1,21 +1,30 @@
+<div class="container">
+    <?php
+    $this->widget('TitleBreadcrumb', [
+        'pageTitle' => 'View Sample #' . $model->id,
+        'breadcrumbItems' => [
+            ['label' => 'Admin', 'href' => '/site/admin'],
+            ['label' => 'Manage', 'href' => '/adminSample/admin'],
+            ['isActive' => true, 'label' => 'View'],
+        ]
+    ]);
+    ?>
 
-<h1>View Sample #<?php echo $model->id; ?></h1>
-<? if (Yii::app()->user->checkAccess('admin')) { ?>
-<div class="actionBar">
-[<?= CHtml::link('Manage Samples', array('admin')) ?>]
-</div>
-<? } ?>
-
-<?php $this->widget('zii.widgets.CDetailView', array(
-	'data'=>$model,
-	'attributes'=>array(
+    <?php $this->widget('zii.widgets.CDetailView', array(
+        'data' => $model,
+        'attributes' => array(
             'id',
             'name',
             'species_id',
             array(
-                'name'=>'attributesList',
-                'type'=>'raw',
-                'value'=> $model->getAttributesList(),
+                'name' => 'attributesList',
+                'type' => 'raw',
+                'value' => $model->getAttributesList(),
             ),
-	),
-)); ?>
+        ),
+        'htmlOptions' => array('class' => 'table table-striped table-bordered dataset-view-table'),
+        'itemCssClass' => array('odd', 'even'),
+        'itemTemplate' => '<tr class="{class}"><th scope="row">{label}</th><td>{value}</td></tr>'
+    )); ?>
+
+</div>