+++ protected/views/funder/view.php
@@ -1,12 +1,25 @@
-<a href="/funder/admin">Back</a>
-<h1>View Funder #<?php echo $model->id; ?></h1>
+<div class="container">
+	<?php
+	$this->widget('TitleBreadcrumb', [
+		'pageTitle' => 'View Funder #' . $model->id,
+		'breadcrumbItems' => [
+			['label' => 'Admin', 'href' => '/site/admin'],
+			['label' => 'Manage', 'href' => '/funder/admin'],
+			['isActive' => true, 'label' => 'View'],
+		]
+	]);
+	$this->widget('zii.widgets.CDetailView', array(
+		'data' => $model,
+		'attributes' => array(
+			'id',
+			'uri',
+			'primary_name_display',
+			'country',
+		),
+		'htmlOptions' => array('class' => 'table table-striped table-bordered dataset-view-table'),
+		'itemCssClass' => array('odd', 'even'),
+		'itemTemplate' => '<tr class="{class}"><th scope="row">{label}</th><td>{value}</td></tr>'
 
-<?php $this->widget('zii.widgets.CDetailView', array(
-	'data'=>$model,
-	'attributes'=>array(
-		'id',
-		'uri',
-		'primary_name_display',
-		'country',
-	),
-)); ?>
+	)); ?>
+
+</div>